+++ test/runtime/frame_test.dart
@@ -10,7 +10,7 @@ void main() {
   setUp(() {
     proto = DarticFuncProto(
       funcId: 0,
-      bytecode: Uint32List.fromList([0, 0, 0]),
+      bytecode: Uint64List.fromList([0, 0, 0]),
       valueRegCount: 4,
       refRegCount: 3,
       paramCount: 2,
